Migrando DatabaseOperationException a server/src
Este cambio mueve \OmegaUp\Exception\DatabaseOperationException al
directorio que puede ser autoloaded.
